Adds python 3.11 and removes python 3.7 due to TensorFlow dropping it.
FYI: Presubmit will likely fail since it is still trying to do python 3.7. Diffbase links the CL to remove python 3.7 testing. Hmm tox (unit test harness) actually just skipped 3.7 (and logged it) since python 3.7 wasn't installed. Not ideal; but helpful in this case.

pybind needed to be updated to support DREAMPlace on Python 3.11. I will send a PR request to them to see if they will fix it in their repo vs. my build scripts.

pygame needed to be pinned to a pre release to support Python 3.11. This is only being done for python 3.11 testing, and is likely to be resolved in week/months. There is a bug listed in the tox.ini. b/267837009

PiperOrigin-RevId: 507463947
Change-Id: Ia30cb687c36b451729e8dc96d7157ef1ab0a3cfc